requirements.txt
neo4j==5.11.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.35.0
numpy==1.25.0
tqdm==4.66.1
pyvis==0.3.1
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv==1.0.0
pydantic>=2.0.0
tabulate>=0.9.0

# LangChain for RAG - Updated to compatible versions
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=0.0.5
langchain-core>=1.0.0
langchain-community>=0.3.0
langsmith>=0.1.0
openai>=1.0.0
google-generativeai>=0.3.0

# Document Parsing - Free Solutions (Primary)
unstructured[pdf]==0.11.8
pillow==10.1.0
pdfminer.six==20221105
python-magic==0.4.27
pytesseract==0.3.10
matplotlib>=3.5.0
unstructured-inference>=0.7.0

# Document Parsing - Optional Free Tools
# marker-pdf==0.2.0  # Uncomment for faster PDF processing
# surya-ocr==0.4.0   # Uncomment for advanced OCR

# Document Parsing - ADE Fallback (Optional)
# landingai==0.3.0   # Uncomment to enable ADE API fallback

